First pass at migrating Ask content
This focuses on simplifying category and answer content.
The categories used for portal topics are broken out as topics for organizing
answers and pages, and answer data is consolidated into single objects that carry
English and Spanish forms of the question, the answer, the short answer and the slug,
as well as the associated topic.

The answer objects can be used to generate wagtail pages, each having a reference to
the source answer.